[forge] Split out test suites from main
There are a lot of tests and cruft in with the main forge cli

Split them out into suites

Test Plan:

running forge on PR succeeds
