+++ testsuite/forge-cli/src/suites/mod.rs
@@ -1,4 +1,15 @@
 // Copyright © Aptos Foundation
+// Parts of the project are originally copyright © Meta Platforms, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod dag;
+pub mod db;
+pub mod indexer;
+pub mod land_blocking;
+pub mod multi_region;
+pub mod netbench;
+pub mod pfn;
+pub mod realistic_environment;
+pub mod realistic_workloads;
+pub mod state_sync;
+pub mod ungrouped;
